+++ VTK/Graphics/vtkImageDataGeometryFilter.cxx
@@ -47,7 +47,7 @@ void vtkImageDataGeometryFilter::Execute()
   vtkCellArray *newPolys=0;
   vtkIdType totPoints, pos;
   int offset[3], numPolys;
-  float x[3];
+  double x[3];
   vtkPointData *pd, *outPD;
   vtkCellData *cd, *outCD;
   vtkImageData *input = this->GetInput();
