+++ VTK/Graphics/vtkVectorNorm.cxx
@@ -37,7 +37,7 @@ void vtkVectorNorm::Execute()
   vtkIdType numVectors, i;
   int computePtScalars=1, computeCellScalars=1;
   vtkFloatArray *newScalars;
-  float v[3], s, maxScalar;
+  double v[3], s, maxScalar;
   vtkDataArray *ptVectors, *cellVectors;
   vtkDataSet *input = this->GetInput();
   vtkDataSet *output = this->GetOutput();
