+++ VTK/Graphics/vtkWarpTo.cxx
@@ -36,11 +36,11 @@ void vtkWarpTo::Execute()
   vtkPoints *newPts;
   vtkIdType ptId, numPts;
   int i;
-  float x[3], newX[3];
+  double x[3], newX[3];
   vtkPointSet *input = this->GetInput();
   vtkPointSet *output = this->GetOutput();
-  float mag;
-  float minMag = 0;
+  double mag;
+  double minMag = 0;
   
   vtkDebugMacro(<<"Warping data to a point");
